Separate metapath-related functions from graph.py
